Refactor sorting methods with trait
