+++ src/Traits/SortableTrait.php
@@ -1,0 +1,110 @@
+<?php
+
+namespace Arrayzy\Traits;
+
+/**
+ * Trait with helpful methods for sorting.
+ *
+ * @author Victor Bocharsky <bocharsky.bw@gmail.com>
+ *
+ * @property array $elements
+ */
+trait SortableTrait
+{
+    /**
+     * {@inheritdoc}
+     *
+     * @link http://php.net/manual/en/function.usort.php
+     */
+    public function customSort(callable $func)
+    {
+        usort($this->elements, $func);
+
+        return $this;
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @link http://php.net/manual/en/function.uksort.php
+     */
+    public function customSortKeys(callable $func)
+    {
+        uksort($this->elements, $func);
+
+        return $this;
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @link http://php.net/manual/en/function.arsort.php
+     * @link http://php.net/manual/en/function.sort.php
+     * @link http://php.net/manual/en/function.asort.php
+     * @link http://php.net/manual/en/function.rsort.php
+     */
+    public function sort($order = SORT_ASC, $strategy = SORT_REGULAR, $preserveKeys = false)
+    {
+        $this->sorting($this->elements, $order, $strategy, $preserveKeys);
+
+        return $this;
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @link http://php.net/manual/en/function.ksort.php
+     * @link http://php.net/manual/en/function.krsort.php
+     */
+    public function sortKeys($order = SORT_ASC, $strategy = SORT_REGULAR)
+    {
+        $this->sortingKeys($this->elements, $order, $strategy);
+
+        return $this;
+    }
+
+    /**
+     * @param array &$elements
+     * @param int $order
+     * @param int $strategy
+     * @param bool $preserveKeys
+     */
+    protected function sorting(array &$elements, $order = SORT_ASC, $strategy = SORT_REGULAR, $preserveKeys = false)
+    {
+        switch ($order) {
+            case SORT_DESC:
+                if ($preserveKeys) {
+                    arsort($elements, $strategy);
+                } else {
+                    rsort($elements, $strategy);
+                }
+                break;
+
+            case SORT_ASC:
+            default:
+                if ($preserveKeys) {
+                    asort($elements, $strategy);
+                } else {
+                    sort($elements, $strategy);
+                }
+        }
+    }
+
+    /**
+     * @param array &$elements
+     * @param int $order
+     * @param int $strategy
+     */
+    protected function sortingKeys(array &$elements, $order = SORT_ASC, $strategy = SORT_REGULAR)
+    {
+        switch ($order) {
+            case SORT_DESC:
+                krsort($elements, $strategy);
+                break;
+
+            case SORT_ASC:
+            default:
+                ksort($elements, $strategy);
+        }
+    }
+}
